api_imageclassification/requirements.txt
Flask
gunicorn
Werkzeug
tensorflow
numpy
pillow
nltk